REFACTOR `ChunkSizer` to `Utils` namespace
